chore(refactor): Prepare for test:
- Refactors components
- Starts testing setup

[Finished]
