+++ package.json
@@ -16,6 +16,7 @@
     "@types/react-dom": "^17.0.9",
     "axios": "^0.26.1",
     "dayjs": "^1.11.0",
+    "enzyme": "^3.11.0",
     "framer-motion": "^6",
     "graphql": "^16.3.0",
     "lodash": "^4.17.21",
@@ -55,8 +56,10 @@
     ]
   },
   "devDependencies": {
+    "@types/enzyme": "^3.10.11",
     "@typescript-eslint/eslint-plugin": "^5.16.0",
     "@typescript-eslint/parser": "^5.16.0",
+    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.7",
     "eslint": "^8.11.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-plugin-import": "^2.25.4",
