+++ src/App.test.tsx
@@ -1,9 +1,14 @@
-import React from 'react';
-import { render, screen } from '@testing-library/react';
-import App from './App';
+import { shallow } from 'enzyme'
+import { findByTestAttr } from 'test/testUtils'
+import App from './App'
 
-test('renders learn react link', () => {
-  render(<App />);
-  const linkElement = screen.getByText(/learn react/i);
-  expect(linkElement).toBeInTheDocument();
-});
+
+const setup = () => {
+  return shallow(<App />);
+}
+
+test('renders without error', () => {
+  const wrapper = setup();
+  const appComponent = findByTestAttr(wrapper, 'component-app');
+  expect(appComponent).toHaveLength(1)
+})
