+++ src/components/ScrollToTop.tsx
@@ -12,3 +12,6 @@ const ScrollToTop = () => {
 };
 
 export default ScrollToTop;
+
+
+
