+++ src/pages/Search/index.tsx
@@ -1,0 +1,13 @@
+import Navigation from './Navigation'
+import SearchControls from './SearchControls';
+
+const Search = () => {
+  return (
+    <>
+      <Navigation />
+      <SearchControls />
+    </>
+  )
+}
+
+export default Search;
