+++ src/routes/protected-routes.tsx
@@ -3,9 +3,9 @@ import { PROTECTED_PATHS } from './constants';
 import { AppRoute } from './types';
 import { lazy } from 'react';
 
-
 const Search = WithSuspense(lazy(() => import('pages/Search')));
 const Result = WithSuspense(lazy(() => import('pages/Result')));
+
 const { SEARCH, RESULT } = PROTECTED_PATHS
 
 const PROTECTED_ROUTES: AppRoute[] = [
