+++ src/setupTests.ts
@@ -3,3 +3,7 @@
 // expect(element).toHaveTextContent(/react/i)
 // learn more: https://github.com/testing-library/jest-dom
 import '@testing-library/jest-dom';
+import EnzymeAdapter from '@wojtekmaj/enzyme-adapter-react-17'
+import Enzyme from 'enzyme'
+
+Enzyme.configure({ adapter: new EnzymeAdapter() });
