+++ src/utils/constant.ts
@@ -1,2 +1,3 @@
 export const AUTH_USER_KEY = "indicina-user-123";
 export const BASE_URL = "https://9uj0ihoex6.execute-api.eu-west-1.amazonaws.com/dev/auth"
+export const GITHUB_LINK = 'https://api.github.com/graphql'
